added goRenderZones

+++ api/file.go
@@ -15,6 +15,7 @@ import (
     //"fmt"
     "io"
     "io/ioutil"
+    "log"
     "os"
     "strings"
 )
@@ -178,7 +179,7 @@ func createFile(fValues *File) error {
         }
 
         if err != nil {
-            removeFile(f)                                                       // need to remove zones/records too !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+            removeFile(f)
             return err
         }
     }
@@ -186,12 +187,8 @@ func createFile(fValues *File) error {
     f.checksum = hex.EncodeToString(checksum[:])
 
     // process data
-    cerr := goScanZones(f, bytes.NewReader(data))
-    err = <-cerr
-    if err != nil {
-        removeFile(f)                                                           // need to remove zones/records too !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-        return err
-    }
+    done := goScanZones(f, bytes.NewReader(data))
+    _ = <-done
 
     return nil
 }
@@ -202,31 +199,40 @@ func readFile(f *File) (file *File, err error) {
     if err != nil {
         return nil, err
     }
+
     checksum := sha1.Sum(data)
     newChecksum := hex.EncodeToString(checksum[:])
 
     if f.checksum != newChecksum {
-        f.checksum = newChecksum
-
         // process data
-        cerr := goScanZones(f, bytes.NewReader(data))
-        err = <-cerr
-        if err != nil {
-            return nil, err
-        }
+        b := bytes.NewBuffer(data)
+        done := goScanZones(f, io.Reader(b))
+        _ = <-done
+
+        f.checksum = newChecksum
     }
 
     return f, nil
 }
 
 func updateFile(f *File, fValues *File) error {
-    data:= []byte("# some updated data")                                        // TBD !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-    checksum := sha1.Sum(data)                                                  // TBD
-    f.checksum = hex.EncodeToString(checksum[:])                                // TBD
+    // collect data
+    b := bytes.NewBuffer([]byte(nil))
+    done := goRenderZones(f, io.Writer(b))
+    _ = <-done
+    data := b.Bytes()
 
-    err := ioutil.WriteFile(f.Path, data, 0644)
-    if err != nil {
-        return err
+    checksum := sha1.Sum(data)
+    newChecksum := hex.EncodeToString(checksum[:])
+
+    if f.checksum !=  newChecksum {
+        // write physical file
+        err := ioutil.WriteFile(f.Path, data, 0644)
+        if err != nil {
+            return err
+        }
+
+        f.checksum = newChecksum
     }
 
     return nil
@@ -243,10 +249,12 @@ func deleteFile(f *File) error {
 
     removeFile(f)   // zeroes f.ID and f.id
 
-    // delete physical file
-    err := os.Remove(path)
-    if err != nil {
-        return err
+    if len(hosts.zoneIndex) == 0 {
+        // delete physical file
+        err := os.Remove(path)
+        if err != nil {
+            return err
+        }
     }
 
     return nil
@@ -254,17 +262,94 @@ func deleteFile(f *File) error {
 
 // -----------------------------------------------------------------------------
 
-func goScanZones(f *File, r io.Reader) chan error {
-    cerr := make(chan error)
+var startMarker string = "##### Start Of Terraform Zone: "
+var endMarker string   = "##### End Of Terraform Zone: "
 
-    go func(cerr chan error) {
-        defer close(cerr)
+// -----------------------------------------------------------------------------
 
-        startMarker := "##### Start Of Terraform Zone: "
-        endMarker   := "##### End Of Terraform Zone: "
+func goRenderZones(f *File, w io.Writer) chan bool {
+    done := make(chan bool)
 
-        // create a hash for the checksum of the file
-//        hash := sha1.New()
+    go func() {
+        defer close(done)
+
+        // Remark:
+        // io.WriteString error cannot happen with the current implementation of updateFile
+        // this would need an io.Writer writing to file instead of to buffer
+        // we leave this for later development
+        // for now we throw a fatal error in case my assumptions are/become wrong
+
+        // render lines for the default external zone
+        zQuery := new(Zone)
+        zQuery.File = f.ID
+        zQuery.Name = "external"
+        z := lookupZone(zQuery)
+
+        lines := goRenderLines(z)
+        for line := range lines {
+            _, err := io.WriteString(w, line + "\n")
+            if err != nil {   // cannot happen
+                log.Fatal(err)
+                // done <- true
+                // return
+            }
+        }
+
+        for _, zone := range hosts.zoneFiles[f.id] {
+            if zone.Name == "external" {
+                continue
+            }
+
+            line := startMarker + zone.Name + " "
+            line = line + strings.Repeat("#", 80 - len(line)) + "\n"
+            _, err := io.WriteString(w, line)
+            if err != nil {   // cannot happen
+                log.Fatal(err)
+                // done <- true
+                // return
+            }
+
+            lines := goRenderLines(zone)
+            for line := range lines {
+                _, err = io.WriteString(w, line + "\n")
+                if err != nil {   // cannot happen
+                    log.Fatal(err)
+                    // done <- true
+                    // return
+                }
+            }
+
+            line = endMarker + zone.Name + " "
+            line = line + strings.Repeat("#", 80 - len(line)) + "\n"
+            _, err = io.WriteString(w, line)
+            if err != nil {   // cannot happen
+                log.Fatal(err)
+                // done <- true
+                // return
+            }
+        }
+
+        // finish goRenderZones()
+        done <- true
+        return
+    }()
+
+    return done
+}
+
+// -----------------------------------------------------------------------------
+
+func goScanZones(f *File, r io.Reader) chan bool {
+    done := make(chan bool)
+
+    go func() {
+        defer close(done)
+
+        // Remark:
+        // Scanner error cannot happen with the current implementation of createFile/readFile
+        // this would need an io.Reader reading from file instead of from buffer
+        // we leave this for later development
+        // for now we throw a fatal error in case my assumptions are/become wrong
 
         // create the default external zone
         zone := "external"
@@ -273,34 +358,25 @@ func goScanZones(f *File, r io.Reader) chan error {
         zValues.File = f.ID
         zValues.Name = zone
         err := createZone(zValues)
-        if err != nil {
-            cerr <- err
-            return
+        if err != nil {   // cannot happen
+            log.Fatal(err)
+            //done <- true
+            //return
         }
         z := lookupZone(zValues)
 
         // create a channel to the external zone
         lines := make(chan string)
-        cerr2 := goScanLines(z, lines)
+        done2 := goScanLines(z, lines)
 
         // save this channel for later use
         linesExternal := lines
-        cerrExternal := cerr2
+        doneExternal := done2
 
         // start scanning
-//        firstLine := true
         scanner := bufio.NewScanner(r)
         for scanner.Scan() {
             line := scanner.Text()
-
-            // update hash
-            // in this case, it is safe ignore errors from io.WriteString
-//            if firstLine {
-//                firstLine = false
-//                _, _ = io.WriteString(hash, line)
-//            } else {
-//                _, _ = io.WriteString(hash, "\n" + line)
-//            }
 
             // scan line
             if lines == linesExternal {
@@ -313,37 +389,31 @@ func goScanZones(f *File, r io.Reader) chan error {
                     zValues.File = f.ID
                     zValues.Name = zone
                     err = createZone(zValues)
-                    if err != nil {
+                    if err != nil {   // cannot happen
                         close(lines)
-                        cerr <- err
-                        return
+                        log.Fatal(err)
+                        //done <- true
+                        //return
                     }
                     z := lookupZone(zValues)
 
                     // create a channel to the new zone
                     lines = make(chan string)
-                    cerr2 = goScanLines(z, lines)
+                    done2 = goScanLines(z, lines)
 
                     continue
                 } 
             } else {
                 if strings.HasPrefix(line, endMarker + zone) {
                     // line is a marker for the end of current zone
-                    // wait for goScanLines(z) of the current zone to finish
+                    // wait for goScanLines() of the current zone to finish
                     close(lines)
-                    err = <-cerr2
-                    if err != nil {
-                        if lines != linesExternal {
-                            close(linesExternal)
-                        }
-                        cerr <- err
-                        return
-                    }
+                    _ = <-done2
 
                     // back to external zone
                     zone = "external"
                     lines = linesExternal
-                    cerr2 = cerrExternal
+                    done2 = doneExternal
 
                     continue
                 }
@@ -352,44 +422,32 @@ func goScanZones(f *File, r io.Reader) chan error {
             // line is not a marker => send to scanLines(z) it
             lines <- line
         }
-        if err := scanner.Err(); err != nil {
+        if err := scanner.Err(); err != nil {   // cannot happen
             // scanner error
-            close(lines)
             if lines != linesExternal {
-                close(linesExternal)
+                close(lines)
             }
-            cerr <- err
-            return
+            close(linesExternal)
+            log.Fatal(err)
+            //done <- true
+            //return
         }
 
         if lines != linesExternal {
             // missing end-marker for current zone => silently ignore
             // wait for goScanLines(z) of the current zone to finish
             close(lines)
-            err = <-cerr2
-            if err != nil {
-                close(linesExternal)
-                cerr <- err
-                return
-            }
+            _ = <-done2
         }
 
         // wait for goScanLines(z) of the external zone to finish
         close(linesExternal)
-        err = <-cerrExternal
-        if err != nil {
-            cerr <- err
-            return
-        }
+        _ = <-doneExternal
 
-        // save checksum of the file
-//        checksum := hash.Sum(nil)
-//        f.checksum = hex.EncodeToString(checksum[:])
-
-        // finish goScanZones(f, r)
-        cerr <- nil
+        // finish goScanZones()
+        done <- true
         return
-    }(cerr)
+    }()
 
-    return cerr
+    return done
 }
